% 
% created by /documair/ at 14_09_12@08:31 
% 
\name{form3split} 
% 
\alias{form3split} 
% 
\title{inverse function of form3join} 
% 
\usage{ 
  form3split(cara,none="-.-",
             OPA="{",CPA="}",
             opa="(",cpa=")",sep="+",
             monitor=rbsa0$monitor$v) 
} 
% 
\examples{ 
  uu <- form3join(letters[1:4]); 
  form3split(uu); 
} 
% 
\details{ 
   
  Of course, it is implicitely supposed that the 
  inversion is unambiguous.\cr 
  The syntax (consistent use of the parentheses and 
  separators) is checked: an error is issued when 
  not correct. The possible \samp{\\n} added by \samp{form3join} 
  is taken into account and removed. 
  For the moment, sep cannot be an empty string. 
   
} 
% 
\description{ 
   
  returns a vector character of the names 
  from a character string generated by \samp{form3join}. 
   
} 
% 
\value{ 
   
  A character vector 
   
} 
% 
 
\arguments{ 
    \item{cara}{Character to be considered.} 
    \item{none}{ idem as in form3join.} 
    \item{OPA}{ idem as in form3join.} 
    \item{CPA}{ idem as in form3join.} 
    \item{opa}{ idem as in form3join.} 
    \item{cpa}{ idem as in form3join.} 
    \item{sep}{ idem as in form3join.} 
    \item{monitor}{ List of monitoring constants, see \samp{rbsa0$monitor$v} to know its structure.} 
} 
 
% 
\keyword{print} 
\keyword{format} 
\keyword{format} 
\keyword{split} 
\seealso{ 
\code{\link{form3join}} 
} 
